# Remove ActiveRecord::OracleEnhanced::Type::Timestamp