--- a/activerecord-oracle_enhanced-adapter.gemspec
+++ b/activerecord-oracle_enhanced-adapter.gemspec
@@ -45,7 +45,6 @@ This adapter is superset of original ActiveRecord Oracle adapter.
     "lib/active_record/connection_adapters/oracle_enhanced/structure_dump.rb",
     "lib/active_record/connection_adapters/oracle_enhanced/version.rb",
     "lib/active_record/oracle_enhanced/type/integer.rb",
-    "lib/active_record/oracle_enhanced/type/timestamp.rb",
     "lib/active_record/oracle_enhanced/type/raw.rb",
     "lib/activerecord-oracle_enhanced-adapter.rb",
     "spec/active_record/connection_adapters/oracle_enhanced_adapter_spec.rb",