… lib/active_record/oracle_enhanced/type/timestamp.rb