Let abstract adapter `initialize_type_map` method handles it.
```ruby
        register_class_with_precision m, %r(datetime)i,  Type::DateTime
        ...
        m.alias_type %r(timestamp)i, 'datetime'
```

This pull request addresses 35 failures and 6 errors of ActiveRecord unit tests.

* Before
```
4631 runs, 12553 assertions, 91 failures, 51 errors, 23 skips
```
* After
```
4631 runs, 12757 assertions, 56 failures, 45 errors, 23 skips
```

Here are each failures and errors fixed by this commit.
```ruby
 21) Failure:
AttributeMethodsTest#test_read_attributes_after_type_cast_on_datetime [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:216]:
--- expected
+++ actual
@@ -1 +1 @@
-Thu, 24 Mar 2011 00:00:00 PDT -07:00
+"2011-03-24"



 22) Failure:
AttributeMethodsTest#test_read_attributes_before_type_cast_on_datetime [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:196]:
Expected: nil
  Actual: "345643456"

 25) Failure:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_in_other_time_zone [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:595]:
--- expected
+++ actual
@@ -1 +1 @@
-#<ActiveSupport::TimeZone:0xXXXXXX @name="Pacific Time (US & Canada)", @utc_offset=nil, @tzinfo=#<TZInfo::DataTimezone: America/Los_Angeles>, @current_period=#<TZInfo::TimezonePeriod: #<TZInfo::TimezoneTransitionDefinition: #<TZInfo::TimeOrDateTime: 1457863200>,#<TZInfo::
TimezoneOffset: -28800,3600,PDT>>,#<TZInfo::TimezoneTransitionDefinition: #<TZInfo::TimeOrDateTime: 1478422800>,#<TZInfo::TimezoneOffset: -28800,0,PST>>>>
+#<ActiveSupport::TimeZone:0xXXXXXX @name="Central Time (US & Canada)", @utc_offset=nil, @tzinfo=#<TZInfo::DataTimezone: America/Chicago>, @current_period=#<TZInfo::TimezonePeriod: #<TZInfo::TimezoneTransitionDefinition: #<TZInfo::TimeOrDateTime: 1457856000>,#<TZInfo::Time
zoneOffset: -21600,3600,CDT>>,#<TZInfo::TimezoneTransitionDefinition: #<TZInfo::TimeOrDateTime: 1478415600>,#<TZInfo::TimezoneOffset: -21600,0,CST>>>>



 26) Failure:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_interprets_time_zone_unaware_string_in_time_zone [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:650]:
--- expected
+++ actual
@@ -1 +1 @@
-Tue, 01 Jan 2008 00:00:00 SST -11:00
+"Tue Jan 01 00:00:00 2008"



 27) Failure:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_to_blank_string_returns_nil [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:639]:
Expected " " to be nil.


 28) Error:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_to_utc:
NoMethodError: undefined method `time_zone' for 2008-01-01 00:00:00 UTC:Time
Did you mean?  in_time_zone
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:583:in `block in test_setting_time_zone_aware_attribute_to_utc'
    /home/yahonda/git/rails/activerecord/test/cases/helper.rb:197:in `in_time_zone'
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:578:in `test_setting_time_zone_aware_attribute_to_utc'


 29) Error:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_with_string:
NoMethodError: undefined method `time_zone' for "2007-12-31 13:00:00 -1100":String
Did you mean?  in_time_zone
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:619:in `block (2 levels) in test_setting_time_zone_aware_attribute_with_string'
    /home/yahonda/git/rails/activerecord/test/cases/helper.rb:197:in `in_time_zone'
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:615:in `block in test_setting_time_zone_aware_attribute_with_string'
    /home/yahonda/git/rails/activesupport/lib/active_support/core_ext/range/each.rb:5:in `each'
    /home/yahonda/git/rails/activesupport/lib/active_support/core_ext/range/each.rb:5:in `each'
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:613:in `test_setting_time_zone_aware_attribute_with_string'


 30) Error:
AttributeMethodsTest#test_setting_time_zone_aware_read_attribute:
NoMethodError: undefined method `time_zone' for 2008-01-01 00:00:00 UTC:Time
Did you mean?  in_time_zone
2.3.0 [/var/tmp]$ more diff.txt 
 21) Failure:
AttributeMethodsTest#test_read_attributes_after_type_cast_on_datetime [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:216]:
--- expected
+++ actual
@@ -1 +1 @@
-Thu, 24 Mar 2011 00:00:00 PDT -07:00
+"2011-03-24"



 22) Failure:
AttributeMethodsTest#test_read_attributes_before_type_cast_on_datetime [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:196]:
Expected: nil
  Actual: "345643456"

 25) Failure:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_in_other_time_zone [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:595]:
--- expected
+++ actual
@@ -1 +1 @@
-#<ActiveSupport::TimeZone:0xXXXXXX @name="Pacific Time (US & Canada)", @utc_offset=nil, @tzinfo=#<TZInfo::DataTimezone: America/Los_Angeles>, @current_period=#<TZInfo::TimezonePeriod: #<TZInfo::TimezoneTransitionDefinition: #<TZInfo::TimeOrDateTime: 1457863200>,#<TZInfo::
TimezoneOffset: -28800,3600,PDT>>,#<TZInfo::TimezoneTransitionDefinition: #<TZInfo::TimeOrDateTime: 1478422800>,#<TZInfo::TimezoneOffset: -28800,0,PST>>>>
+#<ActiveSupport::TimeZone:0xXXXXXX @name="Central Time (US & Canada)", @utc_offset=nil, @tzinfo=#<TZInfo::DataTimezone: America/Chicago>, @current_period=#<TZInfo::TimezonePeriod: #<TZInfo::TimezoneTransitionDefinition: #<TZInfo::TimeOrDateTime: 1457856000>,#<TZInfo::Time
zoneOffset: -21600,3600,CDT>>,#<TZInfo::TimezoneTransitionDefinition: #<TZInfo::TimeOrDateTime: 1478415600>,#<TZInfo::TimezoneOffset: -21600,0,CST>>>>



 26) Failure:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_interprets_time_zone_unaware_string_in_time_zone [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:650]:
--- expected
+++ actual
@@ -1 +1 @@
-Tue, 01 Jan 2008 00:00:00 SST -11:00
+"Tue Jan 01 00:00:00 2008"



 27) Failure:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_to_blank_string_returns_nil [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:639]:
Expected " " to be nil.


 28) Error:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_to_utc:
NoMethodError: undefined method `time_zone' for 2008-01-01 00:00:00 UTC:Time
Did you mean?  in_time_zone
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:583:in `block in test_setting_time_zone_aware_attribute_to_utc'
    /home/yahonda/git/rails/activerecord/test/cases/helper.rb:197:in `in_time_zone'
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:578:in `test_setting_time_zone_aware_attribute_to_utc'


 29) Error:
AttributeMethodsTest#test_setting_time_zone_aware_attribute_with_string:
NoMethodError: undefined method `time_zone' for "2007-12-31 13:00:00 -1100":String
Did you mean?  in_time_zone
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:619:in `block (2 levels) in test_setting_time_zone_aware_attribute_with_string'
    /home/yahonda/git/rails/activerecord/test/cases/helper.rb:197:in `in_time_zone'
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:615:in `block in test_setting_time_zone_aware_attribute_with_string'
    /home/yahonda/git/rails/activesupport/lib/active_support/core_ext/range/each.rb:5:in `each'
    /home/yahonda/git/rails/activesupport/lib/active_support/core_ext/range/each.rb:5:in `each'
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:613:in `test_setting_time_zone_aware_attribute_with_string'


 30) Error:
AttributeMethodsTest#test_setting_time_zone_aware_read_attribute:
NoMethodError: undefined method `time_zone' for 2008-01-01 00:00:00 UTC:Time
Did you mean?  in_time_zone
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:606:in `block in test_setting_time_zone_aware_read_attribute'
    /home/yahonda/git/rails/activerecord/test/cases/helper.rb:197:in `in_time_zone'
    /home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:603:in `test_setting_time_zone_aware_read_attribute'


 32) Failure:
AttributeMethodsTest#test_time_attributes_are_retrieved_in_current_time_zone [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:571]:
Expected 2008-01-01 00:00:00 UTC to be a kind of ActiveSupport::TimeWithZone, not Time.


 33) Failure:
AttributeMethodsTest#test_time_zone_aware_attribute_saved [/home/yahonda/git/rails/activerecord/test/cases/attribute_methods_test.rb:631]:
--- expected
+++ actual
@@ -1 +1 @@
-Mon, 20 Feb 2012 09:00:00 CET +01:00
+2012-02-20 09:00:00 UTC

 43) Failure:
ActiveRecord::CacheKeyTest#test_test_cache_key_format_is_not_too_precise [/home/yahonda/git/rails/activerecord/test/cases/cache_key_test.rb:22]:
--- expected
+++ actual
@@ -1 +1 @@
-"active_record/cache_key_test/cache_mes/10000-20160415210601749423"
+"active_record/cache_key_test/cache_mes/10000-20160415210601749424"


 57) Failure:
ActiveRecord::ConnectionAdapters::TypeLookupTest#test_datetime_types [/home/yahonda/git/rails/activerecord/test/cases/connection_adapters/type_lookup_test.rb:53]:
Expected: :datetime
  Actual: :timestamp


 61) Failure:
DateTimePrecisionTest#test_datetime_data_type_with_precision [/home/yahonda/git/rails/activerecord/test/cases/date_time_precision_test.rb:24]:
Expected: 0
  Actual: nil


 62) Failure:
DateTimePrecisionTest#test_formatting_datetime_according_to_precision [/home/yahonda/git/rails/activerecord/test/cases/date_time_precision_test.rb:59]:
Failed assertion, no message given.


 63) Failure:
DateTimePrecisionTest#test_passing_precision_to_datetime_does_not_set_limit [/home/yahonda/git/rails/activerecord/test/cases/date_time_precision_test.rb:40]:
Expected 4 to be nil.


 64) Failure:
DateTimePrecisionTest#test_schema_dump_includes_datetime_precision [/home/yahonda/git/rails/activerecord/test/cases/date_time_precision_test.rb:72]:
Expected /t\.datetime\s+"created_at",\s+precision: 6,\s+null: false$/ to match "# encoding: UTF-8\n# This file is auto-generated from the current state of the database. Instead\n# of editing this file, please use the migrations feature of Active Record to\n# incrementally 
modify your database, and then regenerate this schema definition.\n#\n# Note that this schema.rb definition is the authoritative source for your\n# database schema. If you need to create the application database on another\n# system, you should be using db:schema:load, not
 running all the migrations\n# from scratch. The latter is a flawed and unsustainable approach (the more migrations\n# you'll amass, the slower it'll run and the greater likelihood for issues).\n#\n# It's strongly recommended that you check this file into your version cont
rol system.\n\nActiveRecord::Schema.define(version: 0) do\n\n  create_table \"foos\", force: :cascade do |t|\n    t.timestamp \"created_at\", limit: 6, null: false\n    t.timestamp \"updated_at\", limit: 6, null: false\n  end\n\nend\n".


 65) Failure:
DateTimePrecisionTest#test_timestamps_helper_with_custom_precision [/home/yahonda/git/rails/activerecord/test/cases/date_time_precision_test.rb:32]:
Expected: 4
  Actual: nil


 66) Failure:
DateTimeTest#test_assign_bad_date_time_with_timezone [/home/yahonda/git/rails/activerecord/test/cases/date_time_test.rb:38]:
Expected "2014-07-01T24:59:59GMT" to be nil.


 67) Failure:
DateTimeTest#test_assign_empty_date_time [/home/yahonda/git/rails/activerecord/test/cases/date_time_test.rb:30]:
Expected "" to be nil.


 69) Failure:
DirtyTest#test_nullable_datetime_not_marked_as_changed_if_new_value_is_blank [/home/yahonda/git/rails/activerecord/test/cases/dirty_test.rb:218]:
Expected "" to be nil.


 70) Failure:
DirtyTest#test_time_attributes_changes_with_time_zone [/home/yahonda/git/rails/activerecord/test/cases/dirty_test.rb:80]:
Expected 2016-04-15 21:06:56 UTC to be a kind of ActiveSupport::TimeWithZone, not Time.

100) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:86]:
--- expected
+++ actual
@@ -1 +1 @@
-2004-06-24 16:24:00 +0000
+{1=>2004, 2=>6, 3=>24, 4=>16, 5=>24, 6=>0}



101) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_will_ignore_date_if_empty [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:175]:
Expected {1=>nil, 2=>nil, 3=>nil, 4=>16, 5=>24} to be nil.


102) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_will_ignore_hour_if_blank [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:165]:
Expected {1=>nil, 2=>nil, 3=>nil, 4=>nil, 5=>12, 6=>2} to be nil.


103) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_will_ignore_hour_if_missing [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:154]:
--- expected
+++ actual
@@ -1 +1 @@
-2004-12-12 00:12:02 +0000
+{1=>2004, 2=>12, 3=>12, 5=>12, 6=>2}



104) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_will_raise_on_big_time_if_missing_date_parts [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:125]:
ActiveRecord::MultiparameterAssignmentErrors expected but nothing was raised.


105) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_empty_seconds [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:275]:
--- expected
+++ actual
@@ -1 +1 @@
-2004-06-24 16:24:00 +0000
+{1=>2004, 2=>6, 3=>24, 4=>16, 5=>24, 6=>nil}



106) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_invalid_time_params [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:102]:
ActiveRecord::MultiparameterAssignmentErrors expected but nothing was raised.


107) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_no_date [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:91]:
ActiveRecord::MultiparameterAssignmentErrors expected but nothing was raised.


108) Error:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_old_date:
ArgumentError: wrong number of arguments (given 1, expected 0)
    /home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:121:in `inspect'
    /home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:121:in `test_multiparameter_attributes_on_time_with_old_date'


109) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_raise_on_small_time_if_missing_date_parts [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:136]:
ActiveRecord::MultiparameterAssignmentErrors expected but nothing was raised.


110) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_seconds_will_ignore_date_if_empty [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:185]:
Expected {1=>nil, 2=>nil, 3=>nil, 4=>16, 5=>12, 6=>2} to be nil.


111) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_skip_time_zone_conversion_for_attributes [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:240]:
--- expected
+++ actual
@@ -1 +1 @@
-2004-06-24 16:24:00 UTC
+{1=>2004, 2=>6, 3=>24, 4=>16, 5=>24, 6=>0}



112) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_time_zone_aware_attributes [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:209]:
--- expected
+++ actual
@@ -1 +1 @@
-2004-06-24 23:24:00 UTC
+{1=>2004, 2=>6, 3=>24, 4=>16, 5=>24, 6=>0}



113) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_time_zone_aware_attributes_false [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:225]:
--- expected
+++ actual
@@ -1 +1 @@
-2004-06-24 16:24:00 +0000
+{1=>2004, 2=>6, 3=>24, 4=>16, 5=>24, 6=>0}



114) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_on_time_with_utc [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:196]:
--- expected
+++ actual
@@ -1 +1 @@
-2004-06-24 16:24:00 UTC
+{1=>2004, 2=>6, 3=>24, 4=>16, 5=>24, 6=>0}



115) Error:
MultiParameterAttributeTest#test_multiparameter_attributes_setting_date_and_time_attribute:
NoMethodError: undefined method `year' for {1=>1952, 2=>3, 3=>11, 4=>13, 5=>55}:Hash
    /home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:301:in `test_multiparameter_attributes_setting_date_and_time_attribute'


116) Error:
MultiParameterAttributeTest#test_multiparameter_attributes_setting_date_attribute:
NoMethodError: undefined method `year' for {1=>1952, 2=>3, 3=>11}:Hash
    /home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:289:in `test_multiparameter_attributes_setting_date_attribute'


117) Failure:
MultiParameterAttributeTest#test_multiparameter_attributes_setting_time_but_not_date_on_date_field [/home/yahonda/git/rails/activerecord/test/cases/multiparameter_attributes_test.rb:309]:
ActiveRecord::MultiparameterAssignmentErrors expected but nothing was raised.


139) Failure:
TransactionTest#test_callback_rollback_in_create [/home/yahonda/git/rails/activerecord/test/cases/transactions_test.rb:279]:
[RuntimeError] exception expected, not
Class: <ActiveRecord::StatementInvalid>
Message: <"OCIError: ORA-01858: a non-numeric character was found where a numeric was expected: INSERT INTO \"TOPICS\" (\"TITLE\", \"AUTHOR_NAME\", \"AUTHOR_EMAIL_ADDRESS\", \"WRITTEN_ON\", \"BONUS_TIME\", \"LAST_READ\", \"CONTENT\", \"APPROVED\", \"CREATED_AT\", \"UPDATED
_AT\", \"ID\") VALUES (:a1, :a2, :a3, :a4, :a5, :a6, :a7, :a8, :a9, :a10, :a11)">
---Backtrace---
stmt.c:243:in oci8lib_230.so
/home/yahonda/.rbenv/versions/2.3.0/lib/ruby/gems/2.3.0/gems/ruby-oci8-2.2.1/lib/oci8/cursor.rb:129:in `exec'
/home/yahonda/git/oracle-enhanced/lib/active_record/connection_adapters/oracle_enhanced/oci_connection.rb:175:in `exec_update'
/home/yahonda/git/oracle-enhanced/lib/active_record/connection_adapters/oracle_enhanced/database_statements.rb:124:in `block in exec_insert'
/home/yahonda/git/rails/activerecord/lib/active_record/connection_adapters/abstract_adapter.rb:527:in `block in log'
/home/yahonda/git/rails/activesupport/lib/active_support/notifications/instrumenter.rb:21:in `instrument'
/home/yahonda/git/rails/activerecord/lib/active_record/connection_adapters/abstract_adapter.rb:521:in `log'
/home/yahonda/git/oracle-enhanced/lib/active_record/connection_adapters/oracle_enhanced_adapter.rb:1170:in `log'
/home/yahonda/git/oracle-enhanced/lib/active_record/connection_adapters/oracle_enhanced/database_statements.rb:110:in `exec_insert'
/home/yahonda/git/rails/activerecord/lib/active_record/connection_adapters/abstract/database_statements.rb:123:in `insert'
/home/yahonda/git/rails/activerecord/lib/active_record/connection_adapters/abstract/query_cache.rb:14:in `insert'
/home/yahonda/git/rails/activerecord/lib/active_record/relation.rb:65:in `insert'
/home/yahonda/git/rails/activerecord/lib/active_record/persistence.rb:554:in `_create_record'
/home/yahonda/git/rails/activerecord/lib/active_record/counter_cache.rb:128:in `_create_record'
…